+++ Ninja/src/Player.h
@@ -6,7 +6,7 @@
 class Player : public MoveableObject {
 public:
 	Player();
-	Player(float xPos, float yPos, float xSpeed, float ySpeed, int id, SDL_Rect bounds, SDL_Texture *sprite);
+	Player(int xPos, int yPos, int xSpeed, int ySpeed, int id, SDL_Rect bounds, SDL_Texture *sprite);
 	virtual ~Player();
 
 	void draw();
